% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boostlscpp}
\alias{boostlscpp}
\title{Function to compute least-squares boosting and the p-value}
\usage{
boostlscpp(y, x, bstop = 50L, nu = 0.1)
}
\arguments{
\item{y}{The n by 1 vector of the dependent variable.}

\item{x}{The n by p matrix of the independent variables.}

\item{bstop}{The number of boosting steps. Defaults to 50.}

\item{nu}{The learning rate. Defaults to 0.1.}
}
\value{
The \code{boostlscpp} function returns a list with the following components:
\item{y}{The n by 1 vector of the dependent variable.}
\item{x}{The n by p matrix of the independent variables.}
\item{bstop}{The number of boosting steps.}
\item{beta}{The p by bstop matrix of parameter estimates in all boosting steps.}
\item{beta0}{The bstop by 1 vector of intercept estimates in all boosting steps.}
\item{se}{The p by bstop matrix of standard errors in all boosting steps.}
\item{pval}{The p by bstop matrix of p-values for all parameter estimates in all
boosting steps. If a p-value is equal to 1, it means the corresponding parameter has
not been updated up to that boosting step.}
\item{step_mat}{A p by bstop matrix that records which variable is updated at each
boosting step.}
\item{aic}{A sequence of AIC values of length \code{bstop}.}
}
\description{
This function computes the LS-Boost estimates, the standard errors, and
the p-values for a cross-section regression model.
}
