\name{boostvar-package}
\alias{boostvar-package}
\docType{package}
\title{
\packageTitle{boostvar}
}
\description{
\packageDescription{boostvar}
}
\details{

The DESCRIPTION file:
This package computes the p-value for the two least-squares
boosting (LS-Boost) algorithms in high-dimensional vector autogression 
described in Huang (2022). In addition to p-values, this package also provides standard errors for LS-Boost estimates in high-dimensional vector autoregressions.

A function is also provided to compute the standard errors and p-values in a cross-section regression.
}
\author{
\packageAuthor{boostvar}

Maintainer: \packageMaintainer{boostvar}
}
\references{
Huang, X. (2022) Boosted p-values for high-dimensional vector autoregression, working paper.
}

\keyword{ vector autoregression, least-squares boosting, p-value }

