% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostvar.R
\name{coef.boostvar}
\alias{coef.boostvar}
\title{Coefficient function for boostvar object}
\usage{
\method{coef}{boostvar}(object, stacked = FALSE)
}
\arguments{
\item{object}{The S3 object returned by the \code{boostvar} function.}

\item{stacked}{A logical variable to indicate the format of the returned coefficients. If
the value is \code{TRUE}, the function will return a \eqn{pd * d} matrix of
all coefficients, matching the shape of the parameter in the equation
\deqn{Y = X \phi + u.} If the value is \code{FALSE}, this function simply return the
the same coefficient as the function \code{boostvar} does.}
}
\description{
This function returns the coefficient of an estimated VAR model.
It can return the coefficient estimates in a stacked matrix format.
}
