% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostvar.R
\name{predict.boostvar}
\alias{predict.boostvar}
\title{Prediction function for the boostvar object}
\usage{
\method{predict}{boostvar}(object, newx, k = NULL)
}
\arguments{
\item{object}{The S3 object returned by the \code{boostvar} function.}

\item{newx}{The data at which the prediction is made. \code{newx} must be a \eqn{n * pd}
matrix, same as the data matrix X in equations (3) and (4) in Huang (2022).}

\item{k}{a boosting step, whose corresponding parameter estimates are used in prediction.}
}
\description{
This function can return the prediction for the entire sequence of parameter
estimates or for parameter estimates at a selected boosting step.
}
